# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Description of the API",
        "title": "WB Author API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
...
        "/v1/create/subscription": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Create a new Okx",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Okx"
                ],
                "summary": "Create new Okx",
                "parameters": [
                    {
                        "description": "Create Okx",
                        "name": "Okx",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/j-ai-trade_modules_okx_model_dto.SubscriptionAddRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Okx"
                        }
                    }
                }
            }
        },
...
        "/v1/delete/okx/{id}": {
            "delete": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Delete a new Okx",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Okx"
                ],
                "summary": "Delete new Okx",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Okx ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Okx"
                        }
                    }
                }
            }
        },
...
        "/v1/get/okx/{id}": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Return Okx",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Okx"
                ],
                "summary": "Get Okx",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Okx ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Okx"
                        }
                    }
                }
            }
        },
...
        "/v2/api-key/create": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Create a new ApiKey",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ApiKey"
                ],
                "summary": "Create new ApiKey",
                "parameters": [
                    {
                        "description": "Create ApiKey",
                        "name": "ApiKey",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.ApiKey"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.ApiKey"
                        }
                    }
                }
            }
        },
        "/v2/api-key/delete": {
            "delete": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Delete a new ApiKey",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ApiKey"
                ],
                "summary": "Delete new ApiKey",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "User UUID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.ApiKey"
                        }
                    }
                }
            }
        },
        "/v2/api-key/get": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Return ApiKey",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ApiKey"
                ],
                "summary": "Get ApiKey",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "User UUID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.ApiKey"
                        }
                    }
                }
            }
        },
        "/v2/api-key/list": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Retrieve a list of ApiKeys based on provided filters and pagination",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ApiKey"
                ],
                "summary": "Get a list of ApiKey",
                "parameters": [
                    {
                        "description": "Get ApiKeys",
                        "name": "ApiKey",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.ApiKeyGetListRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.ApiKey"
                            }
                        }
                    }
                }
            }
        },
        "/v2/api-key/update": {
            "put": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Update ApiKey",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ApiKey"
                ],
                "summary": "Update ApiKey",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "User UUID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update User",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.ApiKeyUpdateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.ApiKey"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "common.PaginationRequest": {
            "type": "object",
            "properties": {
                "index": {
                    "type": "integer"
                },
                "size": {
                    "type": "integer"
                }
            }
        },
...
        "j-ai-trade_modules_okx_model_dto.SubscriptionAddRequest": {
            "type": "object"
        },
...
        "model.ApiKey": {
            "type": "object",
            "properties": {
                "ID": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                }
            }
        },
        "model.ApiKeyGetListRequest": {
            "type": "object",
            "properties": {
                "Pagination": {
                    "$ref": "#/definitions/common.PaginationRequest"
                }
            }
        },
        "model.ApiKeyUpdateRequest": {
            "type": "object"
        },
...
        "model.Okx": {
            "type": "object",
            "properties": {
                "ID": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "Bearer": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}